requirements.txt
python-dotenv==1.0.1
pandas==2.2.3
yfinance==0.2.44